deploy/requirements.txt
pycrypto
apache-libcloud
paramiko
